commit nrgx store

+++ src/app/Services/organization/organization.service.ts
@@ -6,55 +6,19 @@ import { SecurityService } from '../Security/security.service';
 import { environment } from 'src/environments/environment';
 import { Observable } from 'rxjs';
 import { Store } from '@ngrx/store';
-import { AuthState } from 'src/app/Store/auth.reducer';
-import { selectToken } from 'src/app/Store/auth.selectors';
-
+import { AuthState } from 'src/app/Store/Authentication/auth.reducer';
 @Injectable({
   providedIn: 'root'
 })
 export class OrganizationService {
-
-  currentToken:string | null = null;
-  token$!: Observable<string | null>;
   
-  constructor(private api:ApiService, private security: SecurityService, private store: Store<{ auth: AuthState }>) {
-
-    // Selecciona el token desde el store
-    this.token$ = this.store.select(selectToken);
-
-    this.token$.subscribe(token => {
-      
-      console.log('Token: ',token)
-
-      if (token) {
-        
-        this.currentToken = token
-        console.log('Token obtenido:', token);
-      } else {
-        console.log('No hay token disponible.');
-      }
-    });
-    
-  }
-
-  private async getToken() {
-    
-    const token = await this.security.loadToken();
-    if (!token) {
-      console.error('Token no encontrado o inválido.');
-      return null;
-    }
-
-    return token;
-    
-  }
-
+  constructor(private api:ApiService, private security: SecurityService, private store: Store<{ auth: AuthState }>) {}
 
   async getOrganizations(): Promise<any>{
 
-    console.log('curenttoken',this.currentToken)
-    const token = this.currentToken
-    
+    this.security.loadToken();
+    const token = this.security.currentToken
+
     if (!token) {
       console.error('Token no encontrado o inválido.'); 
       return null;
@@ -65,7 +29,9 @@ export class OrganizationService {
     const body = undefined
     
     const response = await this.api.createRequest(endpoint, method, body, token);
-    return response
+    const organizacion = response.data
+
+    return organizacion
 
   }
 
@@ -87,33 +53,33 @@ export class OrganizationService {
     return this.api.createRequest(endpoint, method, body, token);
   }
   */
-  async addOrganization(organization: Organization) {
+  async addOrganization(organization: Organization): Promise<any> {
 
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations`;
     const method = 'POST';
     const body = {organization};
 
-    return this.api.createRequest(endpoint, method, body, token);
-  }
-
-  async updateOrganizationV1(organization: Organization) {
-    
-    const token = await this.security.loadToken();
-
-    const organizationId = organization.id
-
-    const endpoint = `${api_url}/organizations/${organizationId}`;
-    const method = 'PUT';
-    const body = {organization};
-
-    return this.api.createRequest(endpoint, method, body, token);
+    const response = await this.api.createRequest(endpoint, method, body, token);
+    return response
   }
 
   async updateOrganization(organization: Organization) {
     
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations`;
     const method = 'POST';
@@ -124,7 +90,13 @@ export class OrganizationService {
 
   async deleteOrganization(organizationId: number) {
 
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/${organizationId}`;
     const method = 'DELETE';
@@ -135,7 +107,13 @@ export class OrganizationService {
 
   async findOrganization(organizationId: number) {
 
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/${organizationId}`;
     const method = 'GET';
@@ -146,7 +124,13 @@ export class OrganizationService {
 
   async assignUserToOrganization(user: string, organization: number) {
 
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/user/assign/${organization}/${user}`;
 
@@ -158,7 +142,13 @@ export class OrganizationService {
 
   async unassignUserFromOrganization(user: string, organization: number) {
 
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/user/unassign/${organization}/${user}`;
     const method = 'POST';
@@ -170,7 +160,13 @@ export class OrganizationService {
   /*
   async assignUserToOrganizationV1(userId: string, organizationId: number) {
 
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/${organizationId}/users/assign`;
     const method = 'POST';
@@ -181,7 +177,13 @@ export class OrganizationService {
 
   async unassignUserFromOrganizationV1(userId: string, organizationId: number) {
 
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/${organizationId}/users/unassign`;
     const method = 'POST';
@@ -193,7 +195,13 @@ export class OrganizationService {
 
   async assignCategoryToOrganization(organization: number, category: number) {
   
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/category/assign/${organization}/${category}`;
 
@@ -205,7 +213,13 @@ export class OrganizationService {
 
   async unassignCategoryFromOrganization(organization: number, category: number) {
     
-    const token = await this.security.loadToken();
+    this.security.loadToken();
+    const token = this.security.currentToken
+
+    if (!token) {
+      console.error('Token no encontrado o inválido.'); 
+      return null;
+    }
 
     const endpoint = `${api_url}/organizations/category/unassign/${organization}/${category}`;
